flask-server_v2/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
reportlab==3.6.8
markdown2==2.4.8
weasyprint==60.1